Remove tame-index
