+++ .github/workflows/ci.yml
@@ -37,7 +37,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-22.04]
-        features: ["--features targets", "--features prefer-index", null]
+        features: ["--features targets", null]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v3
