+++ Cargo.toml
@@ -19,8 +19,6 @@ rust-version = "1.65.0"
 
 [features]
 default = []
-# Uses the index as the source of truth for what features are available for a crate
-prefer-index = ["dep:tame-index"]
 # Adds support for filtering target specific dependencies
 targets = ["cfg-expr/targets"]
 
@@ -29,8 +27,6 @@ targets = ["cfg-expr/targets"]
 cargo_metadata = "0.17"
 # Used to parse and evaluate cfg() expressions for dependencies
 cfg-expr = "0.15"
-# Allows inspection of the cargo registry index(ices)
-tame-index = { version = "0.4", optional = true, default-features = false }
 # Used to create and traverse graph structures
 petgraph = "0.6"
 # Used for checking version requirements
@@ -44,3 +40,5 @@ insta = "1.21"
 similar-asserts = "1.1"
 # Used to deserialize test files into metadata we can load
 serde_json = "1.0"
+# index metadata retrieval
+tame-index = "0.5"
