+++ src/lib.rs
@@ -42,11 +42,9 @@ mod builder;
 mod errors;
 mod pkgspec;
 
-#[cfg(feature = "prefer-index")]
-pub use builder::index;
 pub use builder::{
     features::{Feature, ParsedFeature},
-    Builder, Cmd, LockOptions, NoneFilter, OnFilter, Scope, Target,
+    index, Builder, Cmd, LockOptions, NoneFilter, OnFilter, Scope, Target,
 };
 pub use errors::Error;
 pub use pkgspec::PkgSpec;
